feat(lfa): allow LFA_PRIME from one CPU at a time
LFA_PRIME is a single-threaded operation that is not pinned
to a specific CPU. The implementation must support calls being
issued from different CPUs, even for several calls to prime
the same component. However, those calls must not happen
concurrently.

This patch introduces a spinlock across the LFA_PRIME call to
support this requirement. In case of concurrent calls, the
LFA_BUSY error code is returned to the caller.

Change-Id: I0574a155ea66b527e26b0dd73272a103e4f936b1
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>