docs: update Ignored Checkpatch Warnings chapter
The warning about lines over 80 characters has been deprecated in kernel
since 2020 [1]. And the checkpatch script used in TF-A-ci-scripts was
then updated in 2021 [2]. Several parts of TF-A still use the 80
character limit, so the Line Length chapter in the file
docs/process/coding-style.rst is kept as-is.

The warning about volatile should not appear if the .checkpatch.conf
is used. Add a sentence about .checkpatch.conf to clarify that.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=bdc48fa11e46f867ea4d75fa59ee87a7f48be144
[2]: https://review.trustedfirmware.org/c/ci/tf-a-ci-scripts/+/11115

Change-Id: Ice83660f90969ef9b9e5f1d5afa2e15c032bfdf1
Signed-off-by: Yann Gautier <yann.gautier@st.com>