fix(allwinner): avoid crash when running without DTB
For determining the PMIC type and the bus its connected to, we use the
DTB that is appended to the U-Boot binary, by looking for it in DRAM.
And while we bail out correctly if we don't find it there, we later try
to use the DTB pointer - without checking.

Add a check in sunxi_pmic_setup(), to only proceed if we have found a
DTB before, and exit early otherwise.

This fixes more experimental setups, where TF-A is run without U-Boot.

Change-Id: I9a7677be057a84fe6bee093d098be758970eec81
Signed-off-by: Andre Przywara <andre.przywara@arm.com>