fix: workaround for CVE-2025-10263
This patch applies the workaround for CVE-2025-10263 to Arm
Trusted Firmware in EL3. The issue can occur when a shareable
TLBI is followed by a DSB instruction, and the workaround is
to insert an additional shareable TLBI followed by a DSB.

This issue is a little bit unique in that if even a single core
in a system is affected, then all cores must perform the
workaround. The build flag defaults to zero but must be set to 1
in the platform makefile of any affected platform.

This first iteration of the patch does not touch the GPT library
code since it is not currently used in any production hardware.

Disclosure: https://developer.arm.com/documentation/112137

Change-Id: Id4c0acee41a40a9fbf780792153e5ef328ba8807
Signed-off-by: John Powell <john.powell@arm.com>
(cherry picked from commit 08fab0f083db753682a12dae3c00eedb20f02620)