fix(build): scan symbols until all are resolved
This small change ensures that unresolved references between object
files are resolved regardless of the order that they are provided to the
linker. This is beneficial for us, because it allows us to append
libraries to variables like `BLx_LIBS` without concern for ordering.

No update is necessary for armlink, which already implements this
behaviour by default.

Change-Id: I2fd2aa04f5911a1d45162c065dfe2a5049b5b14f
Signed-off-by: Chris Kay <chris.kay@arm.com>