feat(cpufeat): include enabled security state scope
This patch adds a security state scope to feature function creation.
The scope defines the set of worlds (Secure, Non-secure, and Realm)
for which a given feature is enabled. This allows finer-grained
control of feature visibility and ensures that ID register are
updated based on feature's enablement status for that world.

Signed-off-by: Arvind Ram Prakash <arvind.ramprakash@arm.com>
Change-Id: Ia4ba4ea7c0b020a398785a28e7ce25b00ab1f4ec