fix(fvp): skip SP discovery through FFA_PARTITION_INFO_GET_REGS
The initialization function implemented for the dummy LSP of FVP port
invokes FFA_PARTITION_INFO_GET_REGS to obtain partition properties of
Secure Partitions managed by SPMC. This happens even before the normal
world is booted.

Hafnium SPMC mistakes this as a FF-A invocation from NWd. As per FF-A
version negotiation protocol, Hafnium locks the version of NWd to v1.3
whereas the NWd never got an opportunity to register its own framework
version.

This patch performs early exit from the helper utility to give NWd
endpoint/Hypervisor an opportunity to register its FF-A version with
SPM. We intentionally do not remove the helper utility as it will be
used in a different patchset for a new anticipated feature.

Change-Id: I54087bd2ad53355afeb024c0e4df6a5ba7ab125a
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>