feat(docs): document PSCI power_state
Add reference to the ARM PSCI spec which provides details about the
power_state.
Document about the EXTENDED_STATE_ID too since some users may
require it, but may not be aware of how to enable it.
Along with this, add couple of lines to describe how TF-A PSCI
framework does some level of sanity validation of this state.

Change-Id: I1f37eab957549c47b500c9592ecf4c9814b6cf2a
Signed-off-by: Dhruva Gole <d-gole@ti.com>