PENDING: feat(ti): Update GPIO clock mux during LPM
During low power mode for the WKUP GPIO to act as a wake up source
the GPIO's clock source needs to be updated to 32K clock. Update the
GPIO clock mux to that during suspend. During resume all clock mux are
saved and restored, so in that step the user configured mux value will
be restored.

Change-Id: If024838045a5e693516b2ed9438a824ceb6fdf2a
Signed-off-by: Sebin Francis <sebin.francis@ti.com>