fix(libc): fix coverity deadcode issue
Comparison checks makes no sense due to the way libc
in TF-A has implemeneted long and long long which are the same
so checking for overflow is deadcode.

Coverity reports -
CID 493664: (#1 of 1): Logically dead code (DEADCODE)

The cutoff and cutlim check is good enough to avoid any overflow.
So remove overflow checks added.

Change-Id: I83e4197e5107bf7c5edc8050bc831b721454573b
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>