fix(cpus): workaround for C1-Pro erratum 3706576
C1-Pro erratum 3706576 is a Cat B erratum that applies to
CPU revisions r0p0 and r1p0, and is fixed in r1p1.

This erratum might cause data corruption when Memory read
effect crossing a 64B boundary, which can be avoided by
setting CPUACTLR2_EL1[37] to 1. Setting this bit is expected
to have a negligible performance impact.

SDEN documentation:
https://developer.arm.com/documentation/SDEN-3273080/1300/?lang=en

Change-Id: Ie427e56c682065bdf82da9b11e71da6383db4e73
Signed-off-by: Xialin Liu <xialin.liu@arm.com>