fix(cpus): workaround for C1-Ultra erratum 3324333
C1-Ultra erratum 3324333 is a Cat B erratum that applies
to revision r0p0, and is fixed in r1p0.

This errata can be avoid by having a speculation barrier
instruction following writes to the SSBS register to
ensure the new value of PSTATE.SSBS affects the subsequent
instructions in the execution stream under speculation.

SDEN documentation:
https://developer.arm.com/documentation/111077/8-0

Change-Id: I203238cbb8561cee683c22a6dbe4742702f82763
Signed-off-by: Xialin Liu <xialin.liu@arm.com>