fix(cpus): workaround for C1-Ultra erratum 3926381
C1-Ultra erratum 3926381 is a Cat B erratum that applies
to revision r1p0 and is open.

This errata can be avoided by executing an implementation
specific instruction patching sequence as soon as possible
after boot. After it is applied, the code only converts
WFx and WFxT instructions to NOP when PSTATE.SM=1 or when
PSTATE.ZA=1.

SDEN documentation:
https://developer.arm.com/documentation/111077/8-0

Change-Id: I2e0f3a715670aaac116c7d3c5f5992ff7ab05ba3
Signed-off-by: Xialin Liu <xialin.liu@arm.com>