fix(cpus): workaround for C1-Premium erratum 3815514
C1-Premium erratum 3815514 is a Cat B erratum that applies
to revisions r0p0, r1p0 and is still open.

The erratum can be avoided by setting CPUACTLR5[13] to 1.
This is expected to result in a small performance degradation
for workloads that use MTE. The degradation might be
approximately 1.6% when using MTE imprecise mode or 0.9% for
MTE precise mode.

SDEN documentation:
https://developer.arm.com/documentation/111078/latest

Change-Id: Id52a1a077459bd16de16b1ae00fc783250d197ed
Signed-off-by: Xialin Liu <xialin.liu@arm.com>