fix(cpus): workaround for C1-Premium erratum 3324333
C1-Premium erratum 3324333 is a Cat B erratum that applies
to revision r0p0, and it is fixed in r1p0.

This errata can be avoided by having a speculation barrier instruction
to ensure the new value of PSTATE.SSBS affects the subsequent
instructions in the execution stream under speculation.

SDEN documentation:
https://developer.arm.com/documentation/111078/latest

Change-Id: Ic9f09c56e7cc94f3d45e86d284971ee2b4b0fb40
Signed-off-by: Xialin Liu <xialin.liu@arm.com>
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>