fix(cpus): workaround for C1-Premium erratum 3926381
C1-Premium erratum 3926381 is a Cat B erratum that applies
to revision r1p0 and is still open.

This errata can be avoided by converting WFx and WFxT
instructions to NOP when PSTATE.SM=1. After it is applied,
the code only converts WFx and WFxT instructions to NOP when
PSTATE.SM=1 or when PSTATE.ZA=1.

SDEN documentation:
https://developer.arm.com/documentation/111078/8-0/?lang=en

Change-Id: I24483fa88c6292f6dbe2950ebef88eebb5cc4e8d
Signed-off-by: Xialin Liu <xialin.liu@arm.com>