fix(cpus): workaround for Cortex-A720AE erratum 3456103
Cortex-A720AE erratum 3456103 is a Cat B erratum that applies
to revisions r0p0 and r0p1, and is still open.

This errata can be avoided by adding a speculation barrier
instruction following writes to the SSBS register to
ensure the new value of PSTATE.SSBS affects the subsequent
instructions in the execution stream under speculation.

SDEN documentation:
https://developer.arm.com/documentation/SDEN-3090091

Change-Id: Ia240a697d8e99bd4fbf4c92720d5228513080088
Signed-off-by: John Powell <john.powell@arm.com>