fix(locks): restore spin_trylock's ability to acquire a lock
The FEAT_STATE conversion of spinlocks (patch
38e580e6411c7a2eb2801a6aacb0a19bb9b1ac46) missed a branch between the
two `mov` instructions and as a result the `mov	rx, #1` is always
followed by `mov rx, #0` causing the spin_trylock() function to always
return failure.

Putting the branch back into the inline assembly is difficult without
extra C control logic that would add extra instructions. So instead
keep only the exclusives in assembly and convert all control logic to C.
This emits code that is functionally equivalent to the original with the
same number of instructions.

Change-Id: If97ea5fd46321ec18c8f92368355b1bc004a7cf0
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>