fix(cpus): workaround for Cortex-X925 erratum 3324334
Cortex-X925 erratum 3324334 is a Cat B erratum that
applies to revisions r0p0, r0p1 and is fixed in r0p2.

This erratum can be avoided by adding a speculation barrier
instruction following writes to the SSBS register to
ensure the new value of PSTATE.SSBS affects the subsequent
instructions in the execution stream under speculation.

SDEN documentation:
https://developer.arm.com/documentation/109180/latest/

Change-Id: I57d2b306b0a3128f3786f4797e6765234ad429cf
Signed-off-by: Xialin Liu <xialin.liu@arm.com>